refactor: move more queries from tb to ch
