+++ apps/billing/src/lib/db-marketing/schemas/serper.ts
@@ -2,8 +2,8 @@ import { relations } from "drizzle-orm";
 import { index, int, json, mysqlTable, text, timestamp, varchar } from "drizzle-orm/mysql-core";
 import { createSelectSchema } from "drizzle-zod";
 import type { z } from "zod";
-import { searchQueries } from "./searchQuery";
 import { firecrawlResponses } from "./firecrawl";
+import { searchQueries } from "./searchQuery";
 
 // Main SearchResponse table
 export const serperSearchResponses = mysqlTable(
