+++ apps/billing/src/trigger/glossary/_generate-glossary-entry.ts
@@ -1,13 +1,13 @@
-import { task } from "@trigger.dev/sdk/v3";
-import { keywordResearchTask } from "./keyword-research";
-import { generateOutlineTask } from "./generate-outline";
-import { draftSectionsTask } from "./draft-sections";
-import { seoMetaTagsTask } from "./seo-meta-tags";
-import { createPrTask } from "./create-pr";
-import { AbortTaskRunError } from "@trigger.dev/sdk/v3";
 import { db } from "@/lib/db-marketing/client";
-import { eq } from "drizzle-orm";
 import { entries } from "@/lib/db-marketing/schemas";
+import { task } from "@trigger.dev/sdk/v3";
+import { AbortTaskRunError } from "@trigger.dev/sdk/v3";
+import { eq } from "drizzle-orm";
+import { createPrTask } from "./create-pr";
+import { draftSectionsTask } from "./draft-sections";
+import { generateOutlineTask } from "./generate-outline";
+import { keywordResearchTask } from "./keyword-research";
+import { seoMetaTagsTask } from "./seo-meta-tags";
 
 export type CacheStrategy = "revalidate" | "stale";
 /**
