+++ internal/clickhouse/src/billing.ts
@@ -1,6 +1,41 @@
 import { z } from "zod";
 import type { Querier } from "./client";
 
+// get the billable ratelimits for a workspace in a specific month.
+// month is not zero-indexed -> January = 1
+export function getBillableRatelimits(ch: Querier) {
+  return async (args: {
+    workspaceId: string;
+    year: number;
+    month: number;
+  }): Promise<number> => {
+    const query = ch.query({
+      query: `
+    SELECT
+      sum(count) as count
+    FROM billing.billable_ratelimits_per_month_v1
+    WHERE workspace_id = {workspaceId: String}
+    AND year = {year: Int64}
+    AND month = {month: Int64}
+    GROUP BY workspace_id, year, month
+    `,
+      params: z.object({
+        workspaceId: z.string(),
+        year: z.number().int(),
+        month: z.number().int().min(1).max(12),
+      }),
+      schema: z.object({
+        count: z.number().int(),
+      }),
+    });
+
+    const res = await query(args);
+    if (!res) {
+      return 0;
+    }
+    return res.at(0)?.count ?? 0;
+  };
+}
 // get the billable verifications for a workspace in a specific month.
 // month is not zero-indexed -> January = 1
 export function getBillableVerifications(ch: Querier) {
@@ -13,7 +48,7 @@ export function getBillableVerifications(ch: Querier) {
       query: `
     SELECT
       sum(count) as count
-    FROM billing.billable_verifications_per_month_v1
+    FROM billing.billable_verifications_per_month_v2
     WHERE workspace_id = {workspaceId: String}
     AND year = {year: Int64}
     AND month = {month: Int64}
